Fix build for examples/
Update Cargo.toml so that examples/rtic.rs can be build. The default
build target is not updated, which means it is still possible to run `cargo
build` successfully on x86_64 by default (without installing the
thumbv7em toolchain target). I was not able to test this change as I do
not have access to this device. Added some documentation to the example
on how to run it.

Disable autoexamples in Cargo.toml as examples/basic.rs contains invalid
syntax.

It is now possible to build the examples using:
$ cargo build --target thumbv7em-none-eabihf --examples