Fix SPI comunication
This fixes #25.

Any try to call the `init` function will fail with an `UnsupportedChip`
error without this (using the defautl `sync` feature).

Previous attempts were made to fix this (#28 and #38) but these where
made before `embedded-hal` `1.0.0` and currently have conflicts
preventing their merge (and can't directly be used in a project using
`embedded-hal` `1.0.0`

closes #28